skyline/requirements.txt
argparse~=1.4.0
beautifulsoup4~=4.7.1
reportlab~=3.5.12
Pillow~=5.3.0